Extracted KairosDBClient interface, added queryMetricTags method.
