+++ test/resources/com/arpnetworking/kairos/client/KairosDbClientImplTest.testCreateRollup.json
