+++ test/resources/com/arpnetworking/kairos/client/KairosDbClientImplTest.testCreateRollup.response.json
