+++ test/resources/com/arpnetworking/kairos/client/KairosDbClientImplTest.testQueryMetric.json
