+++ test/resources/com/arpnetworking/kairos/client/KairosDbClientImplTest.testQueryMetric.response.json
