+++ test/resources/com/arpnetworking/kairos/client/KairosDbClientImplTest.testQueryMetricTags.json
@@ -1,0 +1,5 @@
+{
+  "metrics": [{
+    "name": "metric.name"
+  }]
+}
