+++ test/resources/com/arpnetworking/kairos/client/KairosDbClientImplTest.testRollupsList.response.json
