+++ test/resources/com/arpnetworking/kairos/client/KairosDbClientImplTest.testUpdateRollup.json
