+++ test/resources/com/arpnetworking/kairos/client/KairosDbClientImplTest.testUpdateRollup.response.json
